refactor(elfeed): define feeds in config file
Centralise configuration and setting feeds in config file instead of
doing that in ~/org/elfeed.org.

This avoids the quite annoying bug in which the org-based definition of
feeds appeared in the overview the config file. The bad part is that in
spite of the maintenability increase and bug fix, the code is less
readable (it went from a simple org heading with tags to a list definition
in elisp)